requirements.txt
python-telegram-bot[job-queue]==21.6
psycopg2-binary==2.9.9
python-dotenv==1.0.1
tzdata==2025.2
